Move new trajectory code to control package
