+++ src/main/java/org/frcteam2910/common/control/SplinePathSegment.java
@@ -1,4 +1,4 @@
-package org.frcteam2910.common.control2;
+package org.frcteam2910.common.control;
 
 import org.frcteam2910.common.math.Rotation2;
 import org.frcteam2910.common.math.Vector2;
