+++ src/main/java/org/frcteam2910/common/control/Trajectory.java
@@ -1,244 +1,241 @@
 package org.frcteam2910.common.control;
 
 import org.frcteam2910.common.math.MathUtils;
-import org.frcteam2910.common.math.Rotation2;
-import org.frcteam2910.common.math.Vector2;
-import org.frcteam2910.common.motion.MotionProfile;
-import org.frcteam2910.common.motion.TrapezoidalMotionProfile;
 
-import java.io.Serializable;
+import java.util.ArrayList;
+import java.util.List;
 
-/**
- * A trajectory describes how a path is followed.
- * It contains all the motion profiles for a path.
- */
-public class Trajectory implements Serializable {
-	private static final long serialVersionUID = -7533657200117435410L;
+public class Trajectory {
+    private final Path path;
 
-	private final Path path;
-	private final MotionProfile[] profiles;
-	private Segment[] segments;
-	private double duration = 0.0;
+    private final double duration;
 
-	private double[] maxSegmentVelocities;
-	private double[] maxSegmentAccelerations;
+    private List<ConstrainedPathState> constrainedPathStates = new ArrayList<>();
+    private double[] pathStateStartTimes;
 
-	private double[] profileStartTimes;
+    public Trajectory(Path path, ITrajectoryConstraint[] trajectoryConstraints, double sampleDistance) {
+        this.path = path;
 
-	public Trajectory(Path path, ITrajectoryConstraint... constraints) {
-		this(0.0, 0.0, path, constraints);
-	}
+        double distance = 0.0;
+        ConstrainedPathState lastState = new ConstrainedPathState(
+                path.calculate(distance),
+                0.0,
+                0.0,
+                0.0, // Trajectory starting velocity
+                0.0
+        );
+        while (distance < path.getLength()) {
+            Path.State startingState = path.calculate(distance);
 
-	public Trajectory(double startingVelocity, double endingVelocity, Path path, ITrajectoryConstraint... constraints) {
-		this.path = path;
+            double profileLength = sampleDistance;
+            if (distance + sampleDistance > path.getLength()) {
+                profileLength = path.getLength() - distance;
+            }
 
-		maxSegmentVelocities = new double[path.getSegments().size()];
-		maxSegmentAccelerations = new double[path.getSegments().size()];
+            Path.State endingState = path.calculate(distance + profileLength);
 
-		// First iterate forwards, finding the max velocities when we accelerate
-		for (int i = 0; i < maxSegmentVelocities.length; i++) {
-			PathSegment pathSegment = path.getSegments().get(i);
+            double startingVelocity = lastState.endingVelocity;
 
-			double startVelocity = startingVelocity;
-			if (i > 0) {
-				startVelocity = maxSegmentVelocities[i - 1];
-			}
+            double maxEndingVelocity = Double.POSITIVE_INFINITY;
 
-			double maxVelocity = Double.POSITIVE_INFINITY;
-			for (ITrajectoryConstraint constraint : constraints) {
-				maxVelocity = Math.min(maxVelocity, constraint.getMaxVelocity(pathSegment));
-			}
-			if (maxVelocity < 0.0 || !Double.isFinite(maxVelocity)) {
-				throw new RuntimeException("Illegal max velocity");
-			}
+            for (ITrajectoryConstraint constraint : trajectoryConstraints) {
+                maxEndingVelocity = Math.min(constraint.getMaxVelocity(endingState), maxEndingVelocity);
+            }
 
-			double maxAcceleration = Double.POSITIVE_INFINITY;
-			for (ITrajectoryConstraint constraint : constraints) {
-				maxAcceleration = Math.min(maxAcceleration, constraint.getMaxAcceleration(pathSegment, startVelocity));
-			}
-			if (maxAcceleration < 0.0 || !Double.isFinite(maxAcceleration)) {
-				throw new RuntimeException("Illegal max acceleration");
-			}
+            ConstrainedPathState state = new ConstrainedPathState(
+                    startingState,
+                    profileLength,
+                    startingVelocity,
+                    maxEndingVelocity,
+                    0.0
+            );
 
-			// Now check if we can reach our max velocity
-			double deltaVelocity = maxVelocity - startVelocity;
-			if (deltaVelocity > 0.0) {
-				double accelTime = deltaVelocity / maxAcceleration;
+            // If the max ending velocity is lower than the starting velocity we know that we have to decelerate
+            double maxDeltaVelocity = maxEndingVelocity - startingVelocity;
 
-				double accelDist = 0.5 * maxAcceleration * accelTime * accelTime + startVelocity * accelTime;
+            // Calculate the optimal acceleration for this profile
+            double optimalAcceleration = Math.pow(maxDeltaVelocity, 2.0) / (2.0 * profileLength) + (startingVelocity / profileLength) * maxDeltaVelocity;
+            if (MathUtils.epsilonEquals(optimalAcceleration, 0.0)) {
+                // We are neither accelerating or decelerating
+                state.acceleration = 0.0;
+                state.endingVelocity = state.startingVelocity;
+            } else if (optimalAcceleration > 0.0) {
+                // We are accelerating
+                double maxStartingAcceleration = Double.POSITIVE_INFINITY;
+                double maxEndingAcceleration = Double.POSITIVE_INFINITY;
+                for (ITrajectoryConstraint constraint : trajectoryConstraints) {
+                    maxStartingAcceleration = Math.min(constraint.getMaxAcceleration(startingState, startingVelocity), maxStartingAcceleration);
+                    maxEndingAcceleration = Math.min(constraint. getMaxAcceleration(endingState, startingVelocity), maxEndingAcceleration); // TODO: Use endingVelocity instead of startingVelocity
+                }
 
-				if (accelDist > pathSegment.getLength()) {
-					double[] roots = MathUtils.quadratic(0.5 * maxAcceleration, startVelocity, -pathSegment.getLength());
+                // Take the lower of the two accelerations
+                double acceleration = Math.min(maxStartingAcceleration, maxEndingAcceleration);
 
-					double maxAllowableAccelTime = Math.max(roots[0], roots[1]);
+                // Use the optimal acceleration if we can
+                acceleration = Math.min(acceleration, optimalAcceleration);
 
-					maxVelocity = startVelocity + maxAcceleration * maxAllowableAccelTime;
-				}
-			}
+                // Find the maximum velocity we can reach during this profile
+                double[] roots = MathUtils.quadratic(0.5 * acceleration, startingVelocity, -profileLength);
+                double duration = Math.max(roots[0], roots[1]);
 
-            maxSegmentVelocities[i] = maxVelocity;
-			maxSegmentAccelerations[i] = maxAcceleration;
-		}
+                state.endingVelocity = startingVelocity + acceleration * duration;
+                state.acceleration = acceleration;
+            } else {
+                // If we can decelerate before we reach the end of the profile, use that deceleration.
+                // This acceleration may not be achievable. When we go over the trajectory in reverse we will take care
+                // of this.
+                state.acceleration = optimalAcceleration;
+            }
 
-		for (int i = maxSegmentVelocities.length - 1; i >= 0; i--) {
-			PathSegment pathSegment = path.getSegments().get(i);
+            constrainedPathStates.add(state);
+            lastState = state;
 
-			double endVelocity = endingVelocity;
-			if (i < maxSegmentVelocities.length - 1) {
-				endVelocity = maxSegmentVelocities[i + 1];
-			}
+            distance += profileLength;
+        }
 
-			double maxVelocity = Double.POSITIVE_INFINITY;
-			for (ITrajectoryConstraint constraint : constraints) {
-				maxVelocity = Math.min(maxVelocity, constraint.getMaxVelocity(pathSegment));
-			}
-			if (maxVelocity < 0.0 || !Double.isFinite(maxVelocity)) {
-				throw new RuntimeException("Illegal max velocity");
-			}
+        for (int i = constrainedPathStates.size() - 1; i >= 0; i--) {
+            ConstrainedPathState constrainedState = constrainedPathStates.get(i);
 
-			double maxAcceleration = Double.POSITIVE_INFINITY;
-			for (ITrajectoryConstraint constraint : constraints) {
-				maxAcceleration = Math.min(maxAcceleration, constraint.getMaxAcceleration(pathSegment, endVelocity));
-			}
-			if (maxAcceleration < 0.0 || !Double.isFinite(maxAcceleration)) {
-				throw new RuntimeException("Illegal max acceleration: " + maxAcceleration);
-			}
+            constrainedState.endingVelocity = 0.0; // Trajectory ending velocity
+            if (i != constrainedPathStates.size() - 1) {
+                constrainedState.endingVelocity = constrainedPathStates.get(i + 1).startingVelocity;
+            }
 
-			// Now check if we can reach our max velocity
-			double deltaVelocity = maxVelocity - endVelocity;
-			if (deltaVelocity > 0.0) {
-				double decelTime = deltaVelocity / maxAcceleration;
+            // Check if we are decelerating
+            double deltaVelocity = constrainedState.endingVelocity - constrainedState.startingVelocity;
+            if (deltaVelocity < 0.0) {
+                // Use the deceleration constraint for when we decelerate
+                double deceleration = Double.POSITIVE_INFINITY;
+                for (ITrajectoryConstraint constraint : trajectoryConstraints) {
+                    deceleration = Math.min(deceleration, constraint.getMaxDeceleration(constrainedState.pathState, constrainedState.endingVelocity));
+                }
 
-				double decelDist = 0.5 * maxAcceleration * decelTime * decelTime + endVelocity * decelTime;
+                // Find how long it takes for us to decelerate to the ending velocity
+                double decelTime = deltaVelocity / -deceleration;
 
-				if (decelDist > pathSegment.getLength()) {
-					double[] roots = MathUtils.quadratic(0.5 * maxAcceleration, endVelocity, -pathSegment.getLength());
+                // Find how far we travel while decelerating
+                double decelDist = 0.5 * deceleration * Math.pow(decelTime, 2.0) + constrainedState.endingVelocity * decelTime;
 
-					double maxAllowableDecelTime = Math.max(roots[0], roots[1]);
+                // If we travel too far we have to decrease the starting velocity
+                if (decelDist > constrainedState.length) {
+                    // We can't decelerate in time. Change the starting velocity of the segment so we can.
+                    double[] roots = MathUtils.quadratic(0.5 * deceleration, constrainedState.endingVelocity, -constrainedState.length);
 
-					maxVelocity = endVelocity + maxAcceleration * maxAllowableDecelTime;
-				}
-			}
+                    // Calculate the maximum time that we can decelerate
+                    double maxAllowableDecelTime = Math.max(roots[0], roots[1]);
 
-			maxSegmentVelocities[i] = Math.min(maxSegmentVelocities[i], maxVelocity);
-			maxSegmentAccelerations[i] = Math.max(maxSegmentAccelerations[i], maxAcceleration);
-		}
+                    // Find what are starting velocity can be in order to end at our ending velocity
+                    constrainedState.acceleration = -deceleration;
+                    constrainedState.startingVelocity = constrainedState.endingVelocity + deceleration * maxAllowableDecelTime;
+                }
+            }
+        }
 
-		this.profiles = new MotionProfile[path.getSegments().size()];
-		this.profileStartTimes = new double[profiles.length];
+        pathStateStartTimes = new double[constrainedPathStates.size()];
 
-		MotionProfile.Goal lastPosition = new MotionProfile.Goal(0, startingVelocity);
-		for (int i = 0; i < profiles.length; i++) {
-			// Create the motion constraints for this segment
-			MotionProfile.Constraints segmentConstraints = new MotionProfile.Constraints(maxSegmentVelocities[i], Math.max(maxSegmentAccelerations[i], MathUtils.EPSILON));
+        double duration = 0.0;
+        for (int i = 0; i < constrainedPathStates.size(); i++) {
+            pathStateStartTimes[i] = duration;
+            duration += constrainedPathStates.get(i).getDuration();
+        }
+        this.duration = duration;
+    }
 
-			// Look ahead to see the end velocity for the segment
-			double endVelocity = endingVelocity;
-			if (i < profiles.length - 1) {
-				endVelocity = maxSegmentVelocities[i + 1];
-			}
+    public State calculate(double time) {
+        int start = 0;
+        int end = constrainedPathStates.size() - 1;
+        int mid = start + (end - start) / 2;
+        while (start <= end) {
+            mid = (start + end) / 2;
 
-			MotionProfile.Goal endPosition = new MotionProfile.Goal(path.getDistanceToSegmentEnd(i), endVelocity);
-			profiles[i] = new TrapezoidalMotionProfile(lastPosition, endPosition, segmentConstraints);
+            if (time > pathStateStartTimes[mid] + constrainedPathStates.get(mid).getDuration()) {
+                start = mid + 1;
+            } else if (time < pathStateStartTimes[mid]) {
+                end = mid - 1;
+            } else {
+                break;
+            }
+        }
 
-			profileStartTimes[i] = duration;
-			duration += profiles[i].getDuration();
+        ConstrainedPathState constrainedPathState = constrainedPathStates.get(mid);
+        return constrainedPathState.calculate(time - pathStateStartTimes[mid]);
+    }
 
-			// The profile may not have been able to finish accelerating. We need to manually find the ending velocity
-			lastPosition = new MotionProfile.Goal(
-					path.getDistanceToSegmentEnd(i),
-					profiles[i].calculate(profiles[i].getDuration()).velocity
-			);
-		}
-	}
+    public double getDuration() {
+        return duration;
+    }
 
-	public Segment calculateSegment(double time) {
-		int profileIndex;
-		double profileTime;
+    public Path getPath() {
+        return path;
+    }
 
-		// Binary search to find the correct motion profile
-		{
-			int start = 0;
-			int end = profiles.length - 1;
-			int mid = start + (end - start) / 2;
+    class ConstrainedPathState {
+        public Path.State pathState;
+        public double length;
+        public double startingVelocity;
+        public double endingVelocity;
+        public double acceleration;
 
-			while (start <= end) {
-				// Mid is halfway between start and end
-				mid = (start + end) / 2;
+        public ConstrainedPathState(Path.State pathState, double length, double startingVelocity, double endingVelocity, double acceleration) {
+            this.pathState = pathState;
+            this.length = length;
+            this.startingVelocity = startingVelocity;
+            this.endingVelocity = endingVelocity;
+            this.acceleration = acceleration;
+        }
 
-				if (time > profileStartTimes[mid] + profiles[mid].getDuration()) {
-					// Our time is greater than the end time of the profile, move start to mid and try again
-					start = mid + 1;
-				} else if (time < profileStartTimes[mid]) {
-					// Our time is less than the start time of the profile, move end to mid and try again
-					end = mid - 1;
-				} else {
-					// We are within the start and end times of our profile. This is the profile we want.
-					break;
-				}
-			}
+        public double getDuration() {
+            if (MathUtils.epsilonEquals(acceleration, 0.0)) {
+                return length / startingVelocity;
+            }
 
-			profileIndex = mid;
-			if (profileIndex >= profileStartTimes.length) {
-				return new Segment(profileIndex, time, Vector2.ZERO, Rotation2.ZERO, Rotation2.ZERO, 0.0,
-						0.0, 0.0, 0.0, 0.0);
-			}
+            if (MathUtils.epsilonEquals(endingVelocity, 0.0)) {
+                return (startingVelocity / -acceleration);
+            }
 
-			profileTime = time - profileStartTimes[profileIndex];
-		}
+            double[] roots = MathUtils.quadratic(0.5 * acceleration, startingVelocity, -length);
 
+            if (acceleration > 0.0) {
+                return Math.max(roots[0], roots[1]);
+            } else {
+                return Math.min(roots[0], roots[1]);
+            }
+        }
 
-		MotionProfile.State state = profiles[profileIndex].calculate(profileTime);
-		Vector2 pathPosition = path.getPositionAtDistance(state.position);
-		Rotation2 pathHeading = path.getHeadingAtDistance(state.position);
-		Rotation2 pathRotation = path.getRotationAtDistance(path.getLength() * (time / getDuration()));
+        public State calculate(double time) {
+            time = MathUtils.clamp(time, 0.0, getDuration());
 
-		return new Segment(profileIndex, time, pathPosition, pathHeading, pathRotation, state.position, state.velocity,
-                state.acceleration, maxSegmentVelocities[profileIndex], maxSegmentAccelerations[profileIndex]);
-	}
+            double distance = 0.5 * acceleration * Math.pow(time, 2.0) + startingVelocity * time + pathState.getDistance();
 
-	public void calculateSegments(double dt) {
-		int segmentCount = (int) Math.ceil(getDuration() / dt);
-		segments = new Segment[segmentCount];
+            return new State(
+                    path.calculate(distance),
+                    acceleration * time + startingVelocity,
+                    acceleration
+            );
+        }
+    }
 
-		for (int i = 0; i < segmentCount; i++) {
-			segments[i] = calculateSegment(i * dt);
-		}
-	}
+    public static class State {
+        private final Path.State pathState;
+        private final double velocity;
+        private final double acceleration;
 
-	public Segment[] getSegments() {
-		return segments;
-	}
+        public State(Path.State pathState, double velocity, double acceleration) {
+            this.pathState = pathState;
+            this.velocity = velocity;
+            this.acceleration = acceleration;
+        }
 
-	public double getDuration() {
-		return duration;
-	}
+        public Path.State getPathState() {
+            return pathState;
+        }
 
-	public static class Segment implements Serializable {
-		private static final long serialVersionUID = 1777234890765500009L;
+        public double getVelocity() {
+            return velocity;
+        }
 
-	    public final int pathSegmentIndex;
-		public final double time;
-		public final Vector2 translation;
-		public final Rotation2 heading;
-		public final Rotation2 rotation;
-		public final double position, velocity, acceleration;
-		public final double maxVelocity;
-		public final double maxAcceleration;
-
-		private Segment(int pathSegmentIndex, double time, Vector2 translation, Rotation2 heading, Rotation2 rotation,
-                        double position, double velocity, double acceleration, double maxVelocity,
-                        double maxAcceleration) {
-		    this.pathSegmentIndex = pathSegmentIndex;
-			this.time = time;
-			this.translation = translation;
-			this.heading = heading;
-			this.rotation = rotation;
-			this.position = position;
-			this.velocity = velocity;
-			this.acceleration = acceleration;
-			this.maxVelocity = maxVelocity;
-			this.maxAcceleration = maxAcceleration;
-		}
-	}
+        public double getAcceleration() {
+            return acceleration;
+        }
+    }
 }
