--- src/main/java/org/frcteam2910/common/control2/ITrajectoryConstraint.java
@@ -1,9 +1,0 @@
-package org.frcteam2910.common.control2;
-
-public interface ITrajectoryConstraint {
-    double getMaxVelocity(Path.State state);
-
-    double getMaxAcceleration(Path.State state, double velocity);
-
-    double getMaxDeceleration(Path.State state, double velocity);
-}
