--- src/main/java/org/frcteam2910/common/control2/PathSegment.java
@@ -1,7 +1,0 @@
-package org.frcteam2910.common.control2;
-
-public abstract class PathSegment {
-    public abstract Path.State calculate(double distance);
-
-    public abstract double getLength();
-}
